tools/cli/requirements.txt
typer[all]>=0.9.0
rich>=13.0.0
httpx>=0.25.0
pydantic>=2.0.0
click>=8.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
jinja2>=3.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
